♻️   Abstract api calls to lib
