+++ package.json
@@ -11,6 +11,7 @@
     "leaflet": "^1.7.1",
     "react": "^16.13.1",
     "react-dom": "^16.13.1",
+    "react-helmet": "^6.1.0",
     "react-leaflet": "^3.0.2",
     "react-router-dom": "^5.2.0",
     "styled-components": "^5.1.1"
